Incorporate list and delete functions into graylog_inputs.py Consolidate create_input func procedures
